Engine history delegate API
The goal is to provide our engines with ability to query for browser history
and to record visits and page metadata updates (e.g. title changes).

GeckoView's APIs are still being implemented in https://bugzilla.mozilla.org/show_bug.cgi?id=1494713,
so this works from the implementation notes. WebView's APIs are well known.

History delegate is optional. Absence of it in settings when configuring an engine as akin to
disabling history tracking.

This patch explicitly doesn't provide an implementation of a delegate.