Bug 1497804 - Enable generating specific metrics API at build time
This commit provides a Gradle script that can be applied from the
application's build script in order to generate the specific metrics
API at build time, when parsing the application's metrics.yaml file.
A stub 'parser.py' file is temporarily provided: this will be replaced
by the official glean_parsers package published on PyPi when 1497800
lands.