Closes #1114: Firm up session storage and restoration around new storage format
Current storage format doesn't preserve session order, and snapshot processing
is a bit vague around the various edge cases.

This patch changes the storage format to use an ordered list (vs a set), and firms
up creation, processing and restoration of snapshots. Guiding principles are:
- be explicit,
- be lenient about what we accept, and strict about what we produce.

Storage format version isn't bumped out of simplicity: there are no production
consumers of this (that we know of) that we'll need to migrate, and so let's
do the simple thing this time.