Add an .editorconfig file to fix ktlint
If an .editorconfig file exists in any parent directory of the checkout
directory, ktlint will fail because it uses those settings to determine what
the indentation should be.

This creates an .editorconfig that ktlint will "stop" at, effectively getting
the .editorconfig defaults. It should only be necessary to add settings to this
file if we want to override the defaults.