Update PingMaker.kt to remove format specifier not supported in SDK version 21
SimpleDateFormat, prior to SDK 24 does not support format specifier 'X' to output the time zone offset.  Switched to use format specifier 'Z' which is supported, although slightly different in it's output.  Instead of hh:mm it outputs hhmm for the timezone offset.  In light of this, I had to alter the tests as DateTimeFormatter.ISO_OFFSET_DATE_TIME is strict in requiring the ':' even though the actual ISO8601 spec is not.  This should allow us to fully support the specified minimum SDK version for both android-components, and the reference browser.
Optimize import statements


Remove unused imports


Remove wildcard import


Modify the formatted date string to add the ":" to the offset manually

This will be required by GCP per @fbertsch
Remove unnecessary if check 

Format specifier 'Z' outputs `+0000` and never the `Z` character so the if statement was always true.
Add a note to pings.md explaining the time formatting issue