Issue #2246: Add template to create new gradle module.
My intention is that the developer will follow along with a guide that
has them copy this template and update it for the new module.

Given it took me over an hour to set up a new module by copying over and
changing individual files, I think this will save a lot of time even if
it grows out-dated. However, we've been developing components long
enough that I don't think module design will churn much anymore.