Issue #651: Add PocketTestResource from FFTV.
This was added from FFTV:
  https://github.com/mozilla-mobile/firefox-tv/blob/785501a9eb6c68386b84d0f1104fc5b2f19ed9cb/app/src/test/java/org/mozilla/tv/firefox/TestResource.kt

It was modified slightly: the second item in the test data array had
its `tv_url` and `authors` field restored (from FFTV git history). The
test data data was modified in the FFTV repo so we could test our fallback
behavior. However, the API provided by a-c should be 100% faithful to
the API's returned data and the applications should decide fallback
behavior so these modifications for FFTV are no longer necessary. Also,
it's bad practice to modify the test data with such tests because that test
data could be updated, removing these modifications. Instead, the test
code should generate the test data necessary to replicate and test this
behavior.