Assure engineview is blurred when conducting search
This is important for TalkBack. When the caret in Gecko moves from a
focusable element (like a link) to a non-focusable element (like a text
node in a P), the link is blurred and the document gets focus.

TalkBack then moves the accessibility focus to the input focus, and this
throws the user off of the search result.