1536792: Reinstate disabled MetricsPingSchedulerTests
In order to accomplish this, a window in `Glean.initialize()` where metrics could be recorded before init was complete needed to be closed.  In order to do this, a new `initializing` flag/state was added to allow for uploading of pings while still preventing recording of metrics during initialization.
Add GleanInitStates enum and convert Glean.initialized to enum initState.
Fix a merge conflict