Make sure metric recording happens in the expected order
This changes glean internals so that recording right before
a call to `sendPings` puts the data in that ping. In order
to sanely allow this, `sendPings` now returns a Deferred
which can be used to check if the ping was correctly taken
care of.
This additionally introduces a function to wait for WorkManager
jobs to be queued (in tests) and an `async` facility in Dispatchers.kt