Make Glean mock webserver reply 200 to all requests
`MockWebServer` provides a `QueuedDispatcher` which
requires the consumer to enqueue a response for every
single received request. If no response is found for
a queued request, the test times out in an unclear way.
This changes the tests to use a mockweb server that always
replies 200: if the specific test receives an unexpected
ping, it will fail due to the unexpected content, rather
than timing out in obscure ways.