Enable fetching the 'metrics.yaml' from the AARs in engine-gecko
This stack adds the following mechanics to support GeckoView metrics to
the Glean SDK and engine-gecko-nightly:

- introduces an option to make `sdk_generator.gradle` look for `metrics.yaml`
in AAR files, in addition to locally;
- ensures the `metrics.yaml` from AAR is available for generating the Kotlin APIs;
- changes the script so that the generation task runs even
if no local file is available, as long as the AAR option is enabled;
- it integrates Glean to engine-gecko-nightly as a compileOnly dependency.

**How does this work?**

When the `allowMetricsFromAAR` option is enabled, the Glean SDK build script registers
an artifact transformation class (i.e. `GleanMetricsYamlTransform`) that, as per the
`registerTransform` block, is applied on an "exploded AAR" (unzipped). It looks for the
'metrics.yaml' in the root dir of the exploded AAR. That file is fed as input to the
`generateMetricsSourceFor*` within the `doFirst` block of that task, along with all
the local Glean registry files.