Add a way to record metrics to Glean when Glean is not initialized.
This adds a queueing mechanism to `Dispatchers.API` in order to do a record/playback of events prior to Glean being initialized.  This will help make it so that libraries/components that consume Glean do not need to worry as much about whether Glean is initialized before they record metrics.