Turn Store.broadcastChannel() into Store.channel() and return just a ReceiveChannel.
Broadcast channels have a complicated lifetime since they can have multiple subscribers. This
means a broadcast channel can stay alive longer than there are subscribers. Using a ReceiveChannel,
which can only have one receiver, simplifies this greatly. Also this guarantees that invokeOnClose()
gets invoked when the channel gets closed AND when the receiver is cancelled (from consume()).

On top of that we do not really need the broadcast functionality. As of now we always will only
have a single consumer of the channel. And if that ever changes we can either create multiple
channels or use Kotlin's ReceiveChannel.broadcast() extension method.