fixes #3962 removed duplicates column and made every query a single transaction
added explicit select instead of * select. A * join select returns columns from both tables when column_names are the same. the alternative would be using the "USING" keyword in JOIN, but the name of the column with id is not the same in TabEntity and TabCollectionEntity.

added @Transaction annotation to make query run in a single transaction. TabCollectionWithTabs defines a @Relation  that makes any query that returns one of the POJO's also fetch all it's relations. That makes all fields in the JOIN clause a separate query that could return inconsistent results if the dataset is modified in the meantime.