Part 3: Simplify device management code
This removes ability to automatically poll for device events.
That removal allows for a nice simplification of the two awkward classes
that split the same responsibilities.

Ability to automatically poll in the background is removed because we now
have proper push support (see `feature-push` and friends). Automatic polling
isn't something we'd like to see applications engage in, so let's not enable them.

Manual event polling is still possible via a simplified API call.
Refreshing constellation state is now possible without polling for events.