Closes #4458: Use cached registration token from ADM if available
Since the ADM instance does not live in the app itself, clearing app
data doesn't remove the registration ID for it. This means that we never
initialize the native PushManager with the cached code.

Our fix is straight-forward: if we have a cached registration ID, invoke
the `PushProcessor#onNewToken` and follow the regular flow.