Issue #5376: Lenient and strict methods for determing whether a string is a URL
From only one method to determine whether a string is a url, create
two -- one for a strict check and one for a lenient check. The strict
check will guarantee that the URL precisely conforms to ICANN standards
(e.g., the TLD in the URL is valid). The lenient check will classify any
string that contains ://, ., or : as a URL.