fix feature-pwa manifests db migration
The version 1 schema of the manifests db was modified without increasing
the schema version. This causes existing databases to not get upgraded
and room to throw an exception. The version number has to be increased
and migrations must be provided. Since the modified schema version 1 was
already released, we have to verify both schema variants can be upgraded
to version 2.