Closes #6300: Use a non-common version prefix for local publication
This change makes sure that gradle will not pick a remote version over a
locally published one. We don't expect any available remote versions
to be of the form "0.0.1-somenumber"!

This addresses an issue seen in Fenix - it switched to using a "pinned"
version of A-C, which broke the existing auto-publication logic.
Local changes would be published, but gradle version resolution would
always pick up the pinned version of A-C instead, likely due to how it
alpha-numberically sorted the versions (e.g. it considered 37.0.1-234553
to be newer than 37.0.1-local55554).