For #7788 - Cache request to be reused after storage permission is granted
In the previous flow following storage permission being granted the prompt that
initiated the file picking would be consumed.
Normally, after actually picking a file we would make sure the prompt that
initiated this would be dismissed and following that we would actually do the
upload.
Because this prompt is already consumed following permission being granted the
upload wouldn't succeed and the user would have to redo that upload action.

As such, we need to persist the original prompt even after permission being
granted.