Bump 'settings' migration version to re-run it
We retain Fennec's preferences on disk, so re-running this migration will
make sure that marketing/experiments settings in Fenix correspond to the FHR setting
in Fennec (they'll be turned off if FHR settings is 'off', and left 'on' if FHR is 'on').

For the users running this migration for the first time, it's just a regular migration.

For the users who already ran this migration, we will re-align data collections settings around
their choices in Fennec.