For #7558 - Expose getBookmarkUrlForKeyword(String)
This new method allows to query AS for if a String (containing one or more
words) is set as a keyword for an existing bookmark.

AS does not yet exposes the keyword in the "Bookmark"s objects so users will
not be able to see/add/modify keywords but with this new method they can be
used if they were set on another platform and then the bookmarks were synced
on an Android device.

The searchTerms for which we query bookmarks can have one or more words, with
the first followed by a space character.
If more words are used we'll try to:
- match the first word against a bookmark keyword
- use the rest of the String as a query parameter. Only if the bookmark url
contains "%s" as a placeholder that we can replace.
Otherwise "%s" (case insensitive) will be removed from the bookmark url.