Issue #8791: SearchMiddleware: Add mechanism for migrating legacy data from app
This allows Fenix to pass its existing search engine selection and custom search engines to us and then
we can import them into our storage.

Eventually I expect this code to go away again. It's to be define how many versions we will have to
wait.

So far this does not migrate shown/hidden search engines. This turned out to be more complex since Fenix
saves multiple configurations based on locale and region. So it is not as straight-forward to figure out
which one to use.