For #9043 - Stop leaking the old MotionEvent handled state
How a previous touch was handled was reset when NestedGeckoView would be called
to act for the next touch.
This means the container of NestedGeckoView would get the old status if
querying for it in onInterceptTouchEvent, before NestedGeckoView having the
chance to act upon the new touch.
Resetting the status when the touch is finished ensures any following checks
before NestedGeckoView had the chance to update the status would return the
default UNHANDLED value.