For #9229 - Dismiss SelectLoginPrompt when navigating to another tab
This would ensure GV will trigger this prompt type for the new tab even if
there is (was) already such a prompt shown in the previous tab.
Upon returning to the previous tab which now has the login select prompt
dismissed users just have to focus any login fields for GV to trigger showing
again the login prompt.