For fenix#17451: add Log.Priority.VERBOSE.
Fenix doesn't want to log debug or verbose messages. However, verbose is
currently missing from ac. If the feature is added in the future, we'll
probably unintentionally log verbose messages. To avoid this, I'm adding
it now and will ignore verbose in a follow-up fenix PR.

I didn't add Logger.verbose because it's not needed to fulfill this
regression prevention requirement.