For fenix#17451: add comment warning about Log.Priority.VERBOSE.
The reason we did this is mentioned in
https://github.com/mozilla-mobile/android-components/pull/9451#issuecomment-763747148
as:

> To limit the options to an accepted subset. (We also did not add the
> ASSERT level).

For a satisfying answer, I think we'd need to ask "Why?" again. However,
I don't want to spend more time on this so I didn't follow up on it. This
is why I didn't mention any reasoning for this in the code comment.