For fenix#18142: Fixing regex in WebURLFinder causing jank
The Regex in WebURLFinder would cause backtracking on super long strings. I noticed that the Regex that we have in URLStringUtils (which the comment in
WebURLFinder.kt refers to) is slightly different and doesn't cause the backtracking on super long string, so I decided to just switch it
instead of creating a new regex / using another class (such as Java.net.URL class).