For #7249 - New ads / search telemetry
This is an upstream of the functionality already existing on Fenix so that it
can be used by multiple AC clients.

AdsTelemetry can identify whether there are ads shown in search results in
which case a (Component.FEATURE_SEARCH to SERP_SHOWN_WITH_ADDS) Fact will be
emitted. If AdsMiddleware is installed it will also track when an ad was
clicked for which a (Component.FEATURE_SEARCH to SERP_ADD_CLICKED) Fact will be
emitted.

InContentTelemetry can identify follow-on and organic web searches for which a
(Component.FEATURE_SEARCH to IN_CONTENT_SEARCH) Fact will be emitted.

In both cases the functionality only works for specific providers declared in
BaseSearchTelemetry and will be limited to only crawl websites matching the
ones declared in web extensions' manifest.json.
As such these lists will have to be permanently kept up-to-date.