Remove unnecessary `readScope` from PlacesStorage
All of the 'read' functions that are present in the variants of this
class are already 'suspend'. Declaring a specific dispatcher and
switching to it is harmless and ensures that we'll always end up doing
this work on a dispatchers.io, but all of the consumers are already
expected to invoke 'reader' methods on a worker thread (e.g.
dispatchers.io). Also, there are no special requirements for reading
from places - e.g. it can be done concurrently from different threads.

Contrast this with writing to places - _that_ needs to be sequential,
and so we maintain an `writeScope` to ensure work happens on a
single-threaded dispatcher.

So, the read scope is basically an unnecessary overhead and a tiny bit
of extra complexity we don't need.