Reworking the logins code to work with the new a-s API
The main conceptual change is that we now differentiate between
different types of login data:
  - `LoginEntry` is a login that was entered by the user.  It's not tied
    to any database record
  - `Login` is a login stored in the database.  It has all the LoginEntry
    data, plus guid, timePasswordChanged, etc.
  - `EncryptedLogin` is a Login with the username/password data encrypted

This raises a question: which class should we use when adapting the API?

I used `LoginEntry` for the save login functionality, because I think it
fits better conceptually.  We're saving the data the user just entered
in the form, not updating any particular record.  In fact
SaveLoginDialogFragment will change which record to update as the user
alters the username textbox.

I also used `LoginEntry` for the login picker, solely because the
selection get passed through the GeckoView autocomplete code, which
means it gets converted to/from a GeckoView LoginEntry.  There's no way
to cleanly do with a `Login` because it has too many fields.

One place where this gets weird is when we're listing login entries and
there's duplicated login data because of sync.  Then we're potentially
creating 2 list items with exactly the same data.  But this is just a
reflection of the current UI, which will indeed present 2 items that
look exactly the same for the user to pick between.

The one place I used `Login` was for the autofill feature.  Here the
situation seemed reversed from the GeckoView one.
`AbstractAutofillConfirmActivity` needs to handle a round-trip between
the 3rd party app and it uses the GUID to do that. I don't
really understand this code though, please correct me if I'm doing
something wrong.

At some point it would be nice to use `EncryptedLogin` in places
where we don't actually need the decrypted data.  For example
showing the user that there's an autofill available.  This would
potentially allow us to delay requiring the encryption key.

Change details:

- Replaced the shouldUpdateOrCreateAsync /
  potentialDupesIgnoringUsername logic with something that should be a
  lot simpler.

- Removed `LoginValidationDelegate` and `GeckoLoginStorageDelegate`.
  The regular `LoginsStorage` API is easy enough to use that we don't
  need to wrap it.

- Updated `SaveLoginDialogFragment` and simplified update() method.

- Test updates:
  - Added functions to create fake logins to test with
  - Removed the unittests for service-sync-logins.  At this point all it's
    just wrapper code, so it didn't seem useful to test it.
  - Updated `SaveLoginDialogFragmentTest.kt` and added more test cases.