Ensure we notice and repair a missing or expired push endpoint.
The core problem is that the device constellation's observer is registered
after we've updated the devices, so the code that checks if re-subscribing is
necessary is never hit.

The easiest way of fixing this was to move the `refreshDevices` call to the
account state state-machine, rather than the authentication state-machine,
meaning we can ensure the `onAuthenticated` notification is delivered before
we refresh the devices.

Note that I also removed the special-case for "SEND_TAB" devices - push is
also used for account-related notifications, such as being disconnected.