Upating the `LoginsStorage` API to match the changes from a-s.
Updated `GeckoLoginStorageDelegate` to use the new `addOrUpdate()`
function.

Updated `LoginValidationDelegate` to use the new `findLoginToUpdate()`
function.  Removed the Result.Error class.  It was only returned from
`ensureValidAsync`, which we don't call anymore because `addOrUpdate()`
now does the validation.  This means that we will now throw an error
instead of doing a no-op, hopefully that's okay.  Removed the not
threadsafe warnings.  I think this the login storage code should be
threadsafe in general because we use Arc/Mutex on the rust side

A key difference is that `shouldUpdateOrCreateAsync()` now inputs
`LoginEntry` rather than a `Login`.  The old code had some confusing
bits with how it handled the `guid` field.  For example, we could pass a
`Login` with a `guid` set to the `SaveLoginPrompt`, but depending on
what the final username was and how the dupe-checking turned out, we
could end up updating a login with a different `guid` or creating a new
login altogether.  Using `LoginEntry` hopefully makes the story clearer.

There are a few edge cases where the behavior may change now that we no
longer use a GUID in the dupe-calculation.  I'm pretty sure that it
shouldn't really change things from a user's point of view, but I need
to double-check the use cases here.

`Login.guid` is now non-nullable.  `LoginEntry` should be used for login
data to be saved.

Replaced the field-level documentation in the `service:sync-logins`
README with a note to check out docs from `cconcept-storage`.  The
README was just a copy of the documentation from `LoginsStorage.kt` and
it seems better refer people there rather than duplicate it.

Removed the unit tests.  The main thing they were testing was
the `LoginValidationDelegate` dupe-handling but now that code is very
simple and and there are tests for `findLoginToUpdate()` in a-s
(components/logins/src/login.rs).