Ensure we tell the autopush server about a new FCM token.
This should fix push for FxA.

I also made a number of other push cleanups - in particular,
we are far more aggressive about subscribing - we no longer
do it on login, but instead on every startup. Note that
the Rust component will not hit the server if a subscription
already exists, so this is still relatively fast, and will pick up
missing subscriptions)

This may not fix broken, non-FxA devices - they don't know they need a new
FCM token. It's not even clear if that's a thing though - but either
`renewRegistration` is not necessary, or it is and non FxA push
will remain broken.