Prepend 'http' to URLs in intents whith no schemes
We observed via Sentry that we were hitting a bunch of 'url without
base' exceptions coming out of metadata observation API. One of these
cases where we'll end-up trying to process a URL without a base (e.g.
without a scheme) is when we're processing external intent and see
something like "example.com", instead of "https://example.com".

In our toolbar, we use String.toNormalizedUrl to handle this case.

In this patch, we now do the same for URLs we encounter in external
intents, as well.