Added versioning system for the SQLCipher migration
Store an int that specifies if we've run the SQLCipher migration.  If we
decide to update the logic later on, this int can be used to determine
which version the user has run.

Don't delete the old crypto key.  This gives us the option of trying to
recover the old data in the case of a migration logic error.