Closes #3784: Properly distinguish between source and target of redirects
We were incorrectly handling redirect flags in the onVisited history
delegate.

Specifically, code in the delegate wrapper didn't properly distinguish
between redirect source and target. If a page is a redirect source, that
means it's the page that is redirecting somewhere. If a page is a
target, that means this is another page was redirecting to.

The comments in the code correctly identified this difference, but the
code itself didn't. Perhaps due to potentially misleading names of
redirect constant flags?

Also, this patch removes a test that was attempting to verify that we
wouldn't notify delegates of redirects (meaning, we won't save them).
However, the test itself was configured incorrectly, so even though it
was passing, we were saving the redirects. And further, we actually do
want to save redirects in the most general sense. E.g. to facilitate
correct autocomplete of both src and target in the toolbar.