For #11527 - Use deprecated APIs for immersive mode since they work better
Opened https://issuetracker.google.com/u/2/issues/214012501 for what seems a
bug with the new insets based APIs.
In the meantime we are using the old OnSystemUiVisibilityChangeListener to
know when to restore immersive mode.

Removed the onWindowFocusChangeListener extension property since by having to
offer it through a getter the current implementation would always leak the old
one.
Fenix wasn't using it when this APIs allowed Fenix to pass such a listener and
there was no issue observed so there should be no observable negative impact.