Fix for  #11829: Have the additional_note from the context menu sticky
For #11829: Have the additional_note from the context menu sticky

A note should be displayed below all the options in the context menu when an image is clicked.
Issue - In landscape mode the note is not visible. The user has to scroll down to see the note

Files modified
1. components/feature/contextmenu/build.gradle

Added the following dependency
implementation Dependencies.androidx_constraintlayout

2. components/feature/contextmenu/src/main/res/layout/mozac_feature_contextmenu_dialog.xml

For #11829 : Have the additional_note from the context menu sticky

File modified

mozac_feature_contextmenu_dialog.xml

Added
android:clipToPadding="false" to recyclerview to avoid the content being clipped when scrolling

Co-Authored-By: Mugurell <Mugurell@users.noreply.github.com>