For #11954 - Prefer crash prompt over notification for background child process crashes
With the previous commit making the crash prompt more suitable for
background child process crashes, we should now favour showing it over
the notification. The notification is too hidden from the user,
resulting in fewer crashes being reported than desired.

On recent SDK versions, however, we may not be able to show the prompt
if the crash occurs while the app is in the background. So still use
the notification on those SDK versions.