No issue: Update Nimbus plugin to use version 93.6.0
As part of mozilla-mobile/fenix#26041, we need to ensure we have at
least version 93.6.0 of the plugin downloaded. The reason we are not
using a newer version, even though they are available, is to reduce the
complexity of things that go into the next AC release in order to get
the Fenix experiment launched.

(cherry picked from commit 95c0400e7634de8e17b58aad281669144bed6636)