fixes #11422: add plugin properties for NimbusGradlePlugin and replace appservices token with version
The processResources task copies resources to their target directory and allows them to be processed via various methods built into it. We process the resources for the NimbusGradlePlugin, replacing usages of `@Versions.mozilla_appservices@` with the actual value from the buildSrc generated Versions class. The NimbusGradlePlugin then uses that properties file to retrieve the application services version at runtime without having to have a separate declaration of the version.

https://docs.gradle.org/current/javadoc/org/gradle/language/jvm/tasks/ProcessResources.html