Close #12766: Use seconds as unit when checking sync token expiry
This issue relates to the increased number of re-authorization we perform for
the sync token.  Previously, we've been comparing sync token expiry (in seconds)
with system time (in milliseconds), which is a mistake.

We only uncovered this issue with more tab syncing added that we are checking
the sync token every time we sync.