For #12882: Allow the QR functionality to be queried an resumed.
Clients can query the QRFragment for isScanInProgress and decide on whether to
resume scanning by calling `scan` in a new `QRFeature` instance.
This may be needed if the process is restarted as a followup to the user
updating system permissions for the app which would recreate the component
owning the QRFeature.